[BE] refactor: 불필요한 참조 못하게 DTO 수정
- SimpleGroupCardResponse 원상복귀
- 위의 Response 상속해서 /group/own response에만 적용되는 DTO 생성

- issue (#191)
